Move log_inline from src/include to src/log
