+++ src/log/log_inline.h
